pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rv16-lib"
version = "1.1.0"
description = "A shared Python utility library for the raVioleria16 project."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyYAML",
    "pydantic"
]

[project.optional-dependencies]
core = [
    "PyYAML==6.0.2",
    "pydantic==2.11.9",
    "httpx==0.28.1"
]
redis = [
    "redis>=6.4.0"
]
